Rename command and folders
