+++ extensions/ql-vscode/src/data-extensions-editor/methods-usage/methods-usage-data-provider.ts
